(Added) Basic Composer setup (#3)
* Add composer.json and update .gitignore

A new composer.json file has been added to the project with details on the project, its author, dependencies, and scripts. The .gitignore file has also been updated to ignore the composer.lock file, besides the already ignored files.

* Add multiple steps to PHP Github workflow

This commit adds several steps to the PHP Github workflow including composer validation, installing project dependencies, caching Composer packages, setting up Go, installing osv-scanner, and scanning composer.lock for vulnerabilities. These enhancements ensure stronger code validations and improve the workflow execution speed.

* Update keywords in composer.json

The "keywords" field in composer.json has been updated. "String Manipulation" has been removed and replaced with "xhprof" and "trace". This change reflects the main functionalities and purpose of the project better.

* Update Symfony console dependency version

The composer.json has been updated to allow for Symfony console versions 6 or 7. This change will provide greater compatibility with installations running older versions of Symfony.

* Update PHP version requirements in composer.json

The PHP version requirements have been updated in composer.json to include PHP 8.0, 8.1, and 8.2 along with the existing 7.4. This will ensure the compatibility of our project with the newer PHP versions.

* Update version constraint for enlightn/security-checker

The version constraint for the enlightn/security-checker package in the require-dev section of the composer.json file has been updated. It was previously allowing any version equal to or above 2.0 but is now configured to allow only versions 1.x and 2.x.

* Update Symfony console dependency in composer.json

The Symfony console requirement has been updated to include version 5. We've made this change to ensure broader compatibility across different Symfony console versions.

* Add vulnerability check to PHP Github workflow

New code was added to the Github PHP workflow script to check for security vulnerabilities on successful composer installations. A corresponding command was also added to composer.json to execute this check. Now, the project’s dependencies will be regularly scanned for known security vulnerabilities, enhancing overall application security.

* Update library description in composer.json

The description field in the composer.json file has been updated. The new description provides more details about the library, stating that it's for tracing XHProf performance data in PHP applications.

* Change minimum-stability to dev in composer.json

The minimum-stability in composer.json has been updated from stable to dev. The new setting allows the installation of development packages, enhancing the flexibility for deploying unstable versions. A "prefer-stable" option was also added to prioritize stable versions whenever possible.

* Update minimum PHP version to 8.0

This commit updates the minimum PHP version to 8.0 for both workflow matrix in Github Actions and the composer requirement. This change reflects the team's decision to stop supporting PHP 7.4 and focus on the newer versions of PHP for future development.

* Update Composer install command in GitHub workflow

The commit replaces --no-suggest with --no-interaction in the composer install command within the GitHub workflow. This is because the --no-suggest option is deprecated as of Composer 2.0. Additionally, the --no-interaction flag ensures a non-interactive mode, ideal for continuous integration environments.