(Added) Addition of Lint Checks to the Project (#6)
* Add PHP parallel lint to dev dependencies

The PHP parallel lint and console highlighter packages have been added to the dev dependencies in the composer.json file. These tools allow for syntax checking and looking for problematic patterns in the code. A new "test:lint" command has also been added under scripts.

* Enable parallel linting in GitHub workflow

Resolved the commenting out of parallel linting in the GitHub PHP workflow. The PHP unit tests will now only run if both the vulnerability checks and parallel linting are successful.

* Add lint test to localTest.sh

The update adds a lint test execution command to the localTest.sh script. This change will ensure that all code is checked for potential errors or warnings before the rest of the tests are run.

* Remove version 80 from local test script

This commit excludes version 80 from the PHP versions loop in our local testing script. This is aimed at ensuring the Docker command runs only for the necessary PHP versions, in this case, versions 81, 82, and 83.