Add PHP Mess Detector testing to workflow (#13)
A PHP Mess Detector (PHPMD) scanner was introduced to the testing workflow. This implementation included the addition of a new ruleset file comprising custom checks for the PHP codebase. As a part of this, the local and GitHub test scripts were updated to run the PHPMD tests, and the PHPMD package was added to the composer dependencies.