(Added) Addition of 'xhprof' extension in PHP setup (#22)
* Add xhprof extension to GitHub workflows

The xhprof extension has been added to both the qodana_code_quality and php workflows. This change provides additional profiling capabilities for performance optimization and is reflected in the shivammathur/setup-php step within each workflow file.

* Update PHP extensions list in GitHub workflows

The PHP extensions list in both qodana_code_quality.yml and php.yml workflows have been sorted alphabetically. With this configuration change, it will be easier to read, modify, or update the list of PHP extensions used by GitHub actions workflows. The 'zip' extension was not removed, but moved according to the alphabetical order.